Add files via upload
This section presents the complete SQL workflow, beginning with the creation of SYNONYMs for OLTP tables to streamline accessibility. It then details the development of the dimension tables and the fact table, followed by the ETL process used to populate them with data. Finally, it includes the execution of analytical queries to determine:
1) the total sales value for each production,
2)the monthly sales value for each theatre, and
3) for each theatre, the theatre name along with the clients who recorded the highest spending.